refactor: Move MultiCheckpointManager to kura/checkpoints/multi.py
- Moved MultiCheckpointManager from kura/v1/kura.py to kura/checkpoints/multi.py
- Updated all imports in v1/__init__.py and main __init__.py
- Added list_checkpoints() and delete_checkpoint() methods to CheckpointManager base class
- Updated tests and tutorial to use new import path
- Enhanced tutorial example 6 to demonstrate JSONL + Parquet format combination

This refactor better organizes checkpoint managers by placing them all in the
kura/checkpoints/ directory, making the codebase more maintainable.
